requirements.txt
asgiref==3.3.1
boto3==1.17.44
botocore==1.20.44
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.44.0
django-countries==7.1
django-crispy-forms==1.11.1
django-storages==1.11.1
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python-dotenv==0.16.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
stripe==2.56.0
urllib3==1.26.4
